Module type Mmap_intf.S

To write to/from the mmap, use the sub function to get the subarray of interest. NOTE that because the map may grow when asking for a sub which extends beyond the current length, a remap may occur. Thus, subs should not be reused or stored, since they may refer to an out of date slice for an old mapping.

type t

val of_fd : Unix.file_descr -> t

val length : t -> int

Length of the mmap; should match the fstat size

val sub : t -> off:int -> len:int -> bigstring

Subarray from offset of the given length

val msync : t -> unit

Perform an msync to flush data to disk

val fstat : t -> Unix.stats

val close : t -> unit