Module Mmap__Mmap_intf

Main mmap types

module type S = sig ... end

To write to/from the mmap, use the sub function to get the subarray of interest. NOTE that because the map may grow when asking for a sub which extends beyond the current length, a remap may occur. Thus, subs should not be reused or stored, since they may refer to an out of date slice for an old mapping. For this reason, subs are not thread safe (a new thread may unmap a sub being used by an older thread).