Module Mmap__Mmap_private

Additional support for mmap; we want to have: 1) growable files; 2) ability to track the max used part of the file

type ('a, 'b) t = {

fd : Unix.file_descr;

kind : ('a, 'b) Stdlib.Bigarray.kind;

mutable buf : ('a, 'b, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t;

}

module type S = sig ... end

What we need to use mmap; for large file sizes, we need to be on a 64 bit machine

module Make_1 : functor (S : S) -> sig ... end

module Make_2 : functor (S : S) -> Mmap.Mmap_intf.S

module Export : sig ... end